requirements.txt
alembic==1.7.5
bcrypt==3.2.0
blinker==1.4
cachelib==0.4.1
cffi==1.15.0
click==8.0.3
colorama==0.4.4
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-DotEnv==0.1.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
idna==3.3
importlib-metadata==4.8.2
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
Mako==1.1.6
MarkupSafe==2.0.1
numpy==1.21.4
pandas==1.3.4
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
scikit-learn==1.0.1
scipy==1.7.3
six==1.16.0
SQLAlchemy==1.4.27
threadpoolctl==3.0.0
Werkzeug==2.0.2
WTForms==3.0.0
zipp==3.6.0
